sanders: Enable DTV support for XT1802
* Thanks to @jhenrique09 to let us know XT1802(Brazil) model supports DTV.

+++ configs/check_features.sh
@@ -16,13 +16,13 @@ if [ "$sku" = "XT1806" ]; then
     rm -r /system/app/NfcNci
 fi
 
-#if ! [ "$sku" = "XT1683" ]; then
-#    # Others variants doesn't have DTV support
+if ! [ "$sku" = "XT1802" ]; then
+    # Others variants doesn't have DTV support
     rm /system/etc/permissions/com.motorola.hardware.dtv.xml
     rm /system/etc/permissions/mot_dtv_permissions.xml
     rm /system/lib/libdtvhal.so
     rm /system/lib/libdtvtuner.so
     rm -r /system/priv-app/DTVPlayer
     rm -r /system/priv-app/DTVService
-#fi
+fi
